Fix cleanup on integration tests
If the tests are ignored based on cluster version number, the cleanup
ends up throwing NRE on some tests, so check if ctx is not null before calling
disconnect.

Reviewed-on: http://review.couchbase.org/75839
Tested-by: Subhashni Balakrishnan <b.subhashni@gmail.com>

+++ src/integration/java/com/couchbase/client/java/SubDocumentTest.java
@@ -105,9 +105,10 @@ public class SubDocumentTest {
     }
 
     @AfterClass
-    public static void disconnect() throws InterruptedException {
-        ctx.destroyBucketAndDisconnect();
-        ctx.disconnect();
+    public static void cleanup() throws Exception {
+        if (ctx != null) {
+            ctx.destroyBucketAndDisconnect();
+        }
     }
 
     @Test
