+++ src/integration/java/com/couchbase/client/java/UserManagementTest.java
@@ -57,7 +57,9 @@ public class UserManagementTest {
 
     @AfterClass
     public static void cleanup() {
-        ctx.destroyBucketAndDisconnect();
+        if (ctx != null) {
+            ctx.destroyBucketAndDisconnect();
+        }
     }
 
     @Test
