+++ src/integration/java/com/couchbase/client/java/auth/PasswordAuthenticatorTest.java
@@ -53,9 +53,11 @@ public class PasswordAuthenticatorTest {
     }
 
     @AfterClass
-    public static void tearDown() throws Exception {
-        ctx.clusterManager().removeUser(username);
-        ctx.destroyBucketAndDisconnect();
+    public static void cleanup() throws Exception {
+        if (ctx != null) {
+            ctx.clusterManager().removeUser(username);
+            ctx.destroyBucketAndDisconnect();
+        }
     }
 
     @Test
