+++ src/integration/java/com/couchbase/client/java/rbac/DataServiceUserTest.java
@@ -90,12 +90,14 @@ public class DataServiceUserTest {
     }
 
     @AfterClass
-    public static void tearDown() throws Exception {
-        cluster.disconnect();
-        roCluster.disconnect();
-        ctx.clusterManager().removeUser(username);
-        ctx.clusterManager().removeUser(roUsername);
-        ctx.destroyBucketAndDisconnect();
+    public static void cleanup() throws Exception {
+        if (ctx != null) {
+            cluster.disconnect();
+            roCluster.disconnect();
+            ctx.clusterManager().removeUser(username);
+            ctx.clusterManager().removeUser(roUsername);
+            ctx.destroyBucketAndDisconnect();
+        }
     }
 
     @Test
