+++ src/integration/java/com/couchbase/client/java/rbac/SearchServiceUserTest.java
@@ -76,12 +76,14 @@ public class SearchServiceUserTest {
     }
 
     @AfterClass
-    public static void tearDown() throws Exception {
-        cluster.disconnect();
-        clusterWithNoFtsPerms.disconnect();
-        ctx.clusterManager().removeUser(username);
-        ctx.clusterManager().removeUser(usernameWithNoPerms);
-        ctx.destroyBucketAndDisconnect();
+    public static void cleanup() throws Exception {
+        if (ctx != null) {
+            cluster.disconnect();
+            clusterWithNoFtsPerms.disconnect();
+            ctx.clusterManager().removeUser(username);
+            ctx.clusterManager().removeUser(usernameWithNoPerms);
+            ctx.destroyBucketAndDisconnect();
+        }
     }
 
     @Test
